Fix instructions side bar localization bug
